backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy>=1.10.0
xgboost==2.0.0
lightgbm==4.1.0
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
shap==0.43.0
lime==0.2.0.1
optuna==3.4.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
joblib==1.3.2
imbalanced-learn==0.11.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
requests==2.32.3
